12-26-2015
Changed the main menu to not show the option to continue if there is no saved game,
instead of showing an error message.

First bit of gameplay done. There is now a row of Tribbles that spawn at the bottom of the board, with a spawn animation.
As the row completes it moves up onto the board. Tribbles on the board will randomly show a blink animation.
You can click groups of 2 or more tribbles and they will fly off the board. Also plays a random sfx of a group of "Woot" sounds.
The tribbles don't fall down yet.